backend/services/stt_service.py
import os
from groq import AsyncGroq
from dotenv import load_dotenv

load_dotenv()

client = AsyncGroq(api_key=os.getenv("GROQ_API_KEY"))

async def transcribe_audio(file_path: str):
    if not os.getenv("GROQ_API_KEY"):
        raise ValueError("GROQ_API_KEY is not set in environment variables")
        
    with open(file_path, "rb") as audio_file:
        transcription = await client.audio.transcriptions.create(
            file=(file_path, audio_file.read()),
            model="whisper-large-v3",
            language="en",
            response_format="verbose_json",
        )
    return transcription.text
